Move JSONBodyProducer to utils for sharing
